Prioritize ref lang TMXs in directories with same name as source files
Loading other reference language TMXs are only loaded into a single map
with the language as the key, so if another TMX with the same country
code and locality are loaded in later from a different subdirectory in
tmx2source/, the previous TMX files will no longer be loaded.

This causes issues if you have a central translation memory directory
for multiple projects, and want to specify reference language TMXs for
each project. If each subdirectory for each project contains a reference
language TMX with the same country code and locality, only the latest
TMX to be loaded will work and only one project will have working
language references.

This workaround fix works by prioritizing searching for TMXs in named
subdirectories in the tmx2source/ directory. It assumes that the
subdirectory to search is the same name as the source file being loaded,
so working in a project with a source file named "SBMP_ZH-EN.docx",
reference language TMXs in the directory "tmx2source/SBMP_ZH-EN/" will
be prioritized in loading.

For example, in the case where the translation memory directory has a
"tmx2source/SBMP_ZH-EN/JA.tmx" and a "tmx2source/JA.tmx", only the TMX
from "tmx2source/SMBP_ZH-EN/JA.tmx" will be loaded as the reference for
the JA country code.